Refactored velocity corrector and pressure solve into a separate class
